+++ include/helpers.h
@@ -3,4 +3,4 @@
 #include <algorithm>
 
 bool getYesNoInputWithDefault(bool defaultValue = false, int maxTries = 3, bool hasDefault = true);
-bool getMissingInfo(long int id, std::string& itemVar, char* itemVarName);
+bool getMissingInfo(long long id, std::string& itemVar, char* itemVarName);
